+++ src/commonMain/kotlin/de/atennert/lcarswm/Window.kt
@@ -3,4 +3,4 @@ package de.atennert.lcarswm
 /**
  * POJO for registered windows.
  */
-data class Window(val id: UInt)
+data class Window(val id: ULong)
